feat: prevent like or unlike twice

+++ controllers/like-controller.js
@@ -6,6 +6,13 @@ const likeControler = {
     try {
       const UserId = helpers.getUser(req).id
       const TweetId = req.params.id
+      const like = await Like.findOne({
+        where: {
+          User_id: UserId,
+          Tweet_id: TweetId
+        }
+      })
+      if (like) throw new Error('已經喜歡過了')
       await Like.create({
         UserId,
         TweetId
@@ -19,6 +26,13 @@ const likeControler = {
     try {
       const UserId = helpers.getUser(req).id
       const TweetId = req.params.id
+      const like = await Like.findOne({
+        where: {
+          User_id: UserId,
+          Tweet_id: TweetId
+        }
+      })
+      if (!like) throw new Error('已經不喜歡了')
       await Like.destroy({
         where: {
           UserId,
